requirements.txt
bluepy==1.2.0
paho-mqtt==1.4.0
redis==2.10.6
